Committing changes to prevent loss. Broken!
