+++ tests/unit_test_base.py
@@ -1,0 +1,16 @@
+import pyproctor
+from mock import Mock, MagicMock
+
+
+class UnitTestBase(pyproctor.TestBase):
+    def setUp(self):
+        self.mock_container()
+
+    def mock_container(self):
+        request = Mock()
+        self.container = Mock()
+        self.container.request = request
+        self.storage = Mock()
+        self.storage.__exit__ = MagicMock(return_value=False)
+        self.storage.__enter__ = MagicMock(return_value=self.storage)
+        self.container.create_master_bucket_storage = MagicMock(return_value=self.storage)
